--- Classes_OOP/LetterStats
@@ -1,26 +1,0 @@
-letter|count_all|count_uppercase|percentage
-a|535|96|8.21
-b|89|0|1.37
-c|196|90|3.01
-d|217|27|3.33
-e|824|97|12.65
-f|178|46|2.73
-g|67|24|1.03
-h|215|0|3.30
-i|500|67|7.68
-j|13|9|0.20
-k|49|22|0.75
-l|277|58|4.25
-m|159|14|2.44
-n|355|82|5.45
-o|435|49|6.68
-p|195|105|2.99
-r|454|19|6.97
-s|532|95|8.17
-t|564|41|8.66
-u|206|4|3.16
-v|66|4|1.01
-w|233|47|3.58
-x|24|0|0.37
-y|130|25|2.00
-z|1|0|0.02
