--- Classes_OOP/WordCount.csv
@@ -1,106 +1,0 @@
-on-105
-published-81
-a-76
-news-73
-city-67
-is-62
-for-47
-weather-46
-forecast-46
-it-46
-some-27
-another-26
-privat-25
-ad-24
-expires-24
-days-24
-left-24
-from-24
-file-24
-los-23
-angeles-23
-temperature-23
-advice-23
-warm-23
-outside-23
-great-23
-time-23
-walk-23
-without-23
-outerwear-23
-part-19
-new-16
-york-16
-selling-16
-the-15
-this-15
-sample-15
-car-15
-sunny-14
-and-12
-second-10
-of-9
-last-9
-advertisement-8
-vilnius-5
-minsk-4
-to-3
-by-2
-city’s-2
-dresden-2
-not-2
-no-2
-more-1
-than-1
-lovers-1
-showed-1
-their-1
-love-1
-tattooing-1
-trolleybus-1
-mythical-1
-iron-1
-wolf-1
-angel-1
-lively-1
-pink-1
-soup-1
-three-1
-armed-1
-robot-1
-conductor-1
-makes-1
-debut-1
-in-1
-german-1
-sinfoniker-1
-says-1
-aim-1
-replace-1
-humans-1
-but-1
-play-1
-music-1
-human-1
-conductors-1
-would-1
-find-1
-impossible-1
-citizens-1
-were-1
-surprised-1
-seimas-1
-elections-1
-something-1
-publish-1
-one-1
-so-1
-interesting-1
-newt-1
-cool-1
-nsflnv-1
-ks-1
-lv-1
-krakow-1
-aaaaaaaaaaaaaa-1
-jjjjjjjjjjjjj-1
