+++ LetterStats
@@ -1,0 +1,26 @@
+letter|count_all|count_uppercase|percentage
+a|585|104|8.24
+b|98|0|1.38
+c|228|99|3.21
+d|238|29|3.35
+e|894|105|12.59
+f|190|51|2.68
+g|73|26|1.03
+h|232|1|3.27
+i|544|71|7.66
+j|14|9|0.20
+k|52|23|0.73
+l|300|62|4.23
+m|168|14|2.37
+n|395|89|5.56
+o|474|52|6.68
+p|211|113|2.97
+r|495|22|6.97
+s|574|101|8.08
+t|615|45|8.66
+u|225|4|3.17
+v|71|4|1.00
+w|248|50|3.49
+x|28|0|0.39
+y|146|26|2.06
+z|2|0|0.03
