+++ NewsFeed.txt
@@ -499,3 +499,41 @@ City: KRAKOW
 AAAAAaaAAAAaAA JJJJjJjJJjJjJ
 Published on: 2024-11-03
 
+--------------------News
+City: NEW YORK
+Local community center opens new library section.
+Published on: 2024-11-10
+
+--------------------Privat Ad
+Get your car insurance at a discount this month!
+Published on: 2024-11-10
+Expires on: 2024-12-31
+Days left: 51
+
+--------------------Weather Forecast
+City: LOS ANGELES
+Forecast for: 2024-11-16
+Temperature: 25°C
+Advice: It is warm outside. It is a great time for a walk without outerwear.
+Expect a sunny day with a light breeze.
+Published on: 2024-11-10
+
+--------------------News
+City: SAN FRANCISCO
+New coffee shop in downtown area attracts many visitors.
+Published on: 2024-11-10
+
+--------------------Privat Ad
+Huge discounts on electronics! Sale ends soon.
+Published on: 2024-11-10
+Expires on: 2024-11-20
+Days left: 10
+
+--------------------Weather Forecast
+City: SEATTLE
+Forecast for: 2024-11-19
+Temperature: 8°C
+Advice: It is chilly. Consider wearing a jacket.
+Rain expected later in the day, carry an umbrella.
+Published on: 2024-11-10
+
