+++ WordCount.csv
@@ -1,0 +1,149 @@
+on-114
+published-87
+a-82
+news-75
+city-71
+is-65
+for-50
+forecast-50
+it-49
+weather-48
+privat-27
+some-27
+ad-26
+expires-26
+days-26
+left-26
+another-26
+temperature-25
+advice-25
+los-24
+angeles-24
+warm-24
+outside-24
+great-24
+time-24
+walk-24
+without-24
+outerwear-24
+from-24
+file-24
+new-19
+part-19
+york-17
+the-16
+this-16
+selling-16
+car-16
+sample-15
+sunny-15
+and-12
+second-10
+of-9
+last-9
+advertisement-8
+vilnius-5
+minsk-4
+in-3
+to-3
+by-2
+city’s-2
+dresden-2
+not-2
+no-2
+day-2
+more-1
+than-1
+lovers-1
+showed-1
+their-1
+love-1
+tattooing-1
+trolleybus-1
+mythical-1
+iron-1
+wolf-1
+angel-1
+lively-1
+pink-1
+soup-1
+three-1
+armed-1
+robot-1
+conductor-1
+makes-1
+debut-1
+german-1
+sinfoniker-1
+says-1
+aim-1
+replace-1
+humans-1
+but-1
+play-1
+music-1
+human-1
+conductors-1
+would-1
+find-1
+impossible-1
+citizens-1
+were-1
+surprised-1
+seimas-1
+elections-1
+something-1
+publish-1
+one-1
+so-1
+interesting-1
+newt-1
+cool-1
+nsflnv-1
+ks-1
+lv-1
+krakow-1
+aaaaaaaaaaaaaa-1
+jjjjjjjjjjjjj-1
+local-1
+community-1
+center-1
+opens-1
+library-1
+section-1
+get-1
+your-1
+insurance-1
+at-1
+discount-1
+month-1
+expect-1
+with-1
+light-1
+breeze-1
+san-1
+francisco-1
+coffee-1
+shop-1
+downtown-1
+area-1
+attracts-1
+many-1
+visitors-1
+huge-1
+discounts-1
+electronics-1
+sale-1
+ends-1
+soon-1
+seattle-1
+chilly-1
+consider-1
+wearing-1
+jacket-1
+rain-1
+expected-1
+later-1
+carry-1
+an-1
+umbrella-1
